backend/requirements.txt
click==8.1.3
sqlalchemy==2.0.15
psycopg2==2.9.5
sqlalchemy-utils==0.39.0
fastapi[email]==0.89.1
fastapi-login==1.8.3
bcrypt==4.0.1
python-multipart==0.0.5
httpx==0.23.3
httpcore==0.16.3
cloud-sql-python-connector[pg8000]
google-cloud-storage
email-validator
